build: Test builds on Travis CI
Compiling mozjs consumes time and resources. This Travis CI configuration
file will build mozjs, try the interpreter, and link with a minimal C++
program on Linux and macOS, using Clang, and additionally on Linux with
GCC 6 and 7 as the compiler. This allows us to catch errors more promptly
when trying patches for mozjs or new versions of mozjs.